[JENKINS-46028] Use primary job as fallback
Currently, test results are only taken from the last builds of this job. The
result is that the first build(s) of a newly created branch (when using multi-
branch jobs) do not run tests in parallel, unless the build for this branch is
non-FAILURE. With short-lived feature branches, this results in unsatisfying
long build times.

In case of multi-branch jobs, other branches could act as a source for test
records as well. The question, which branch to use as reference can be answered
with the SCM's help - the PrimaryInstanceMetadataAction is assigned by several
SCM plugins and denotes one/multiple branches as "primary" (in Git: the default
branch when checking out).